Suggested Changes
config files, master_file_example, and fullMPMU are just stylistic
changes
preprocess was a small bug in the butterworth filter, adding in an
option for bandpass
mexGPU was the windows solution that I found. would need other feedback
to find out if this is the 'correct' option
rez2phy was a small bug
plot_waveforms2 is what I use to look at things before phy.